1.1.18

+++ scripts/effective-tray.mjs
@@ -45,7 +45,7 @@ export class effectiveTray {
   /**
    * Make the effects tray effective
    * @param {ChatMessage5e} message The message on which the tray resides.
-   * @param {HTMLElement} html HTML contents of the message.
+   * @param {HTMLElement} html      HTML contents of the message.
    * Methods lacking documentation below share these parameters.
    */
   static async _effectTray(message, html) {
@@ -81,10 +81,13 @@ export class effectiveTray {
         "EFFECTIVETRAY.TOOLTIP.EffectsApplyTokensLegacy" :
         "EFFECTIVETRAY.TOOLTIP.EffectsApplyTokens"
     ) : "DND5E.EffectsApplyTokens";
-    const lvl = message.flags?.dnd5e?.use?.spellLevel;
+    let spellLevel;
+    if (!message.flags?.dnd5e?.use?.spellLevel) spellLevel = 0;
+    else spellLevel = parseInt(message.flags?.dnd5e?.use?.spellLevel) || null;
+    const effectData = { "flags.dnd5e.spellLevel": spellLevel };
+    const concentration = actor.effects.get(message.getFlag("dnd5e", "use.concentrationId"));
+    const caster = actor.uuid;
     for (const effect of effects) {
-      const concentration = actor.effects.get(message.getFlag("dnd5e", "use.concentrationId"));
-      const caster = actor.uuid;
       const label = effect.duration.duration ? effect.duration.label : "";
       const contents = `
         <li class="effect" data-uuid=${uuid}.ActiveEffect.${effect.id} data-transferred=${effect.transfer}>
@@ -101,29 +104,26 @@ export class effectiveTray {
       tray.querySelector('ul.effects.unlist').insertAdjacentHTML("beforeend", contents);
 
       // Handle click events
-      tray.querySelector(`li[data-uuid="${uuid}.ActiveEffect.${effect.id}"]`)?.querySelector("button").addEventListener('click', async() => {
+      tray.querySelector(`li[data-uuid="${uuid}.ActiveEffect.${effect.id}"]`)?.querySelector("button").addEventListener('click', async () => {
         const mode = tray.querySelector(`[aria-pressed="true"]`)?.dataset?.mode;
         if (!mode || mode === "selected") {
           const actors = new Set();
           for (const token of canvas.tokens.controlled) if (token.actor) actors.add(token.actor);
+          _checkTray(tray);
           for (const actor of actors) {
-            await _applyEffects(actor, effect, lvl, concentration);
-            if (!game.settings.get(MODULE, "dontCloseOnPress")) {
-              tray.classList.add("collapsed");
-              tray.classList.remove("et-uncollapsed");
-            };  
+            await _applyEffects(actor, effect, { effectData, concentration });
           };
         } else {
-          _effectApplicationHandler(tray, effect, lvl, concentration, caster);
+          _effectApplicationHandler(tray, effect, { effectData, concentration, caster });
         };
       });
 
       // Handle legacy targeting mode
       if (game.settings.get(MODULE, "allowTarget") && game.settings.get(MODULE, "contextTarget")) {
-        tray.querySelector(`li[data-uuid="${uuid}.ActiveEffect.${effect.id}"]`)?.querySelector("button").addEventListener('contextmenu', async function(event) {
+        tray.querySelector(`li[data-uuid="${uuid}.ActiveEffect.${effect.id}"]`)?.querySelector("button").addEventListener('contextmenu', async function (event) {
           event.stopPropagation();
           event.preventDefault();
-          _effectApplicationHandler(tray, effect, lvl, concentration, caster);
+          _effectApplicationHandler(tray, effect, { effectData, concentration, caster });
         });
       };
     };
@@ -141,7 +141,7 @@ export class effectiveTray {
       event.preventDefault();
       if (html.querySelector(".et-uncollapsed")) {
         tray.classList.remove("et-uncollapsed");
-      };  
+      };
       if (!html.querySelector(".effects-tray.collapsed")) {
         tray.classList.add("collapsed");
       } else if (html.querySelector(".effects-tray.collapsed")) tray.classList.remove("collapsed");
@@ -177,7 +177,7 @@ export class effectiveTray {
     if (game.settings.get(MODULE, "dontCloseOnPress")) {
       const buttons = tray.querySelectorAll("button");
       for (const button of buttons) {
-        button.addEventListener('click', async() => {
+        button.addEventListener('click', async () => {
           if (!tray.querySelector(".et-uncollapsed")) {
             await tray.classList.add("et-uncollapsed");
             await new Promise(r => setTimeout(r, 108));
@@ -211,7 +211,7 @@ export class effectiveTray {
     const tray = html.querySelector('.effects-tray');
     if (!tray) return;
     const mid = message.id;
-    await tray.classList.add("et-uncollapsed");
+    if (game.settings.get(MODULE, "dontCloseOnPress")) tray.classList.add("et-uncollapsed");
     await new Promise(r => setTimeout(r, 108));
     await tray.classList.remove("collapsed");
     if (game.settings.get(MODULE, "scrollOnExpand")) _scroll(mid);
@@ -246,7 +246,7 @@ export class effectiveDamage {
   /**
    * Make the damage tray effective
    * @param {ChatMessage5e} message The message on which the tray resides.
-   * @param {HTMLElement} html HTML contents of the message.
+   * @param {HTMLElement} html      HTML contents of the message.
    * Methods lacking documentation below share these parameters.
    */
   static _damageTray(message, html) {
@@ -355,10 +355,11 @@ export class effectiveSocket {
 async function _effectSocket(request) {
   if (game.user !== game.users.activeGM) return;
   const targets = request.data.targets;
-  const effect = await fromUuid(request.data.origin);
+  const origin = request.data.origin;
+  const effect = foundry.utils.getType(origin) === "string" ? await fromUuid(origin) : origin;
   const c = await fromUuid(request.data.caster);
   const concentration = c?.effects?.get(request.data.con);
-  const lvl = request.data.lvl;
+  const effectData = request.data.effectData;
   const actors = new Set();
   for (const target of targets) {
     const token = await fromUuid(target);
@@ -366,7 +367,7 @@ async function _effectSocket(request) {
     if (target) actors.add(targetActor);
   };
   for (const actor of actors) {
-    await _applyEffects(actor, effect, lvl, concentration);
+    await _applyEffects(actor, effect, { effectData, concentration });
   };
 };
 
@@ -396,87 +397,90 @@ async function _damageSocket(request) {
 
 /**
  * Handle applying effects to targets: handle it if you can handle it, else make a request via socket
- * @param {HTMLElement} tray HTML element that composes the collapsible tray.
- * @param {ActiveEffect5e} effect The effect to create.
- * @param {number} lvl The spellLevel of the spell the effect is on, if it is on a spell.
+ * @param {HTMLElement} tray             HTML element that composes the collapsible tray.
+ * @param {ActiveEffect5e} effect        The effect to create.
+ * @param {object} effectData            A generic data object that contains spellLevel in a `dnd5e` scoped flag, and whatever else.
  * @param {ActiveEffect5e} concentration The concentration effect on which `effect` is dependent, if it requires concentration.
- * @param {string} caster The Uuid of the actor which originally cast the spell requiring concentration.
+ * @param {string} caster                The Uuid of the actor which originally cast the spell requiring concentration.
  */
-async function _effectApplicationHandler(tray, effect, lvl, concentration, caster) {
+async function _effectApplicationHandler(tray, effect, { effectData, concentration, caster }) {
   if (!game.user.targets.size) return ui.notifications.info(game.i18n.localize("EFFECTIVETRAY.NOTIFICATION.NoTarget"));
-  const targets = Array.from(game.user.targets).map(i => i.document.uuid)
   if (game.user.isGM) {
     const actors = new Set();
     for (const token of game.user.targets) if (token.actor) actors.add(token.actor);
+    _checkTray(tray);
     for (const actor of actors) {
-      await _applyEffects(actor, effect, lvl, concentration);
-      if (!game.settings.get(MODULE, "dontCloseOnPress")) {
-        tray.classList.add("collapsed");
-        tray.classList.remove("et-uncollapsed");
-      };
+      await _applyEffects(actor, effect, { effectData, concentration });
     };
   } else {
+    const [owned, targets] = partitionTargets(game.user.targets);
+    const actors = new Set();
+    for (const token of owned) if (token.actor) actors.add(token.actor);
+    _checkTray(tray);
+    for (const actor of actors) {
+      await _applyEffects(actor, effect, { effectData, concentration });
+    };
+    if (!targets.length) return;
     if (!game.users.activeGM) return ui.notifications.warn(game.i18n.localize("EFFECTIVETRAY.NOTIFICATION.NoActiveGMEffect"));
     const origin = effect.uuid;
     const con = concentration?.id;
-    await game.socket.emit(socketID, { type: "effect", data: { origin, targets, lvl, con, caster } });
-    if (!game.settings.get(MODULE, "dontCloseOnPress")) {
-      tray.classList.add("collapsed");
-      tray.classList.remove("et-uncollapsed");
-    };
+    await game.socket.emit(socketID, { type: "effect", data: { origin, targets, effectData, con, caster } });
   };
 };
 
 /**
  * Apply effect, or refresh its duration (and level) if it exists
- * @param {Actor5e} actor The actor to create the effect on.
- * @param {ActiveEffect5e} effect The effect to create.
- * @param {number} lvl The spellLevel of the spell the effect is on, if it is on a spell.
+ * @param {Actor5e} actor                The actor to create the effect on.
+ * @param {ActiveEffect5e} effect        The effect to create.
+ * @param {object} effectData            A generic data object that contains spellLevel in a `dnd5e` scoped flag, and whatever else.
  * @param {ActiveEffect5e} concentration The concentration effect on which `effect` is dependent, if it requires concentration.
  */
-async function _applyEffects(actor, effect, lvl, concentration) {
+export async function _applyEffects(actor, effect, { effectData, concentration }) {
+
+  // Call the pre effect hook; returning `false` will terminate the function
+  const preCallback = Hooks.call("effectiv.preApplyEffect", actor, effect, { effectData, concentration });
+  if (!preCallback) return;
+
+  // Enable an existing effect on the target if it originated from this effect
   const origin = game.settings.get(MODULE, "multipleConcentrationEffects") ? effect : concentration ?? effect;
   const existingEffect = game.settings.get(MODULE, "multipleConcentrationEffects") ?
     actor.effects.find(e => e.origin === effect.uuid) :
     actor.effects.find(e => e.origin === origin.uuid);
-  let flags;
-  if (game.settings.get(MODULE, "flagLevel") && effect?.parent?.type === "spell") {
-    flags = foundry.utils.deepClone(effect.flags);
-    foundry.utils.mergeObject(flags, {
-      dnd5e: {
-        spellLevel: lvl
-      }
-    });
-  } else flags = effect.flags;
+
   if (existingEffect) {
     if (!game.settings.get(MODULE, "deleteInstead")) {
-      return existingEffect.update({
+      return existingEffect.update(foundry.utils.mergeObject({
         ...effect.constructor.getInitialDuration(),
-        disabled: false,
-        flags: flags
-      });
+        disabled: false
+      }, effectData));
+
+      // Or delete it instead
     } else existingEffect.delete();
-    if ( !game.user.isGM && concentration && !concentration.actor?.isOwner ) {
-      throw new Error(game.i18n.localize("DND5E.EffectApplyWarningConcentration"));
-    }
   } else {
-    const effectData = foundry.utils.mergeObject(effect.toObject(), {
+
+    // Otherwise, create a new effect on the target
+    effect instanceof ActiveEffect ? effect = effect.toObject() : effect;
+    effectData = foundry.utils.mergeObject({
+      ...effect,
       disabled: false,
       transfer: false,
-      origin: origin.uuid,
-      flags: flags
-    });
+      origin: origin.uuid
+    }, effectData);
     const applied = await ActiveEffect.implementation.create(effectData, { parent: actor });
     if (concentration) await concentration.addDependent(applied);
+
+    // Call the effect hook
+    Hooks.callAll("effectiv.applyEffect", actor, effect, { effectData, concentration });
+
     return applied;
   };
 };
 
 /**
  * Apply damage
- * @param {string} id The id of the actor to apply damage to.
+ * @param {string} id      The id of the actor to apply damage to.
  * @param {object} options The options provided by the tray, primarily the multiplier.
- * @param {array} damage An array of objects with the damage type and value that also contain Sets with damage properties.
+ * @param {array} damage   An array of objects with the damage type and value that also contain Sets with damage properties.
  */
 async function _applyTargetDamage(id, options, damage) {
   const actor = fromUuidSync(id);
@@ -496,5 +500,30 @@ async function _scroll(mid) {
   if (window.ui.sidebar.popouts.chat && window.ui.sidebar.popouts.chat.isAtBottom) {
     await new Promise(r => setTimeout(r, 256));
     window.ui.sidebar.popouts.chat.scrollBottom();
+  };
+};
+
+/**
+ * Sort tokens into owned and unowned categories.
+ * @param {Set|array} targets The set or array of tokens to be sorted.
+ * @returns {array}           An Array of length two whose elements are the partitioned pieces of the original
+ */
+export function partitionTargets(targets) {
+  const result = targets.reduce((acc, t) => {
+    if (t.isOwner) acc[0].push(t);
+    else acc[1].push(t.document.uuid);
+    return acc;
+  }, [[], []]);
+  return result;
+};
+
+/**
+ * Sort tokens into owned and unowned categories.
+ * @param {HTMLElement} tray The tray to be collapsed or not collapsed
+ */
+function _checkTray(tray) {
+  if (!game.settings.get(MODULE, "dontCloseOnPress")) {
+    tray.classList.add("collapsed");
+    tray.classList.remove("et-uncollapsed");
   };
 };
